PICPAY feat: Implementado consulta por id

+++ src/main/java/com/br/wes/pay/application/contracts/TransactionGateway.java
@@ -10,5 +10,5 @@ public interface TransactionGateway {
 
 
     Transaction save(Transaction transaction);
-    Optional<Transaction> findById(Long id);
+    Optional<Transaction> findById(Integer id);
 }
